[Refactor] 레포지토리 경로 변경
