--- src/main/java/com/appcenter/marketplace/domain/member/MemberRepository.java
@@ -1,7 +1,0 @@
-package com.appcenter.marketplace.domain.member;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-}
