+++ src/main/java/com/appcenter/marketplace/domain/member/repository/MemberRepository.java
@@ -1,0 +1,8 @@
+package com.appcenter.marketplace.domain.member.repository;
+
+import com.appcenter.marketplace.domain.member.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long>, MemberRepositoryCustom {
+
+}
